+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Rename `widget::RenderIndicator` to `widget::OrIndicators`
 - Change Highlight Symbol to actually be drawn (fixes #1)
   - The Highlight Symbol draw width is configurable
+  - The Highlight Symbol draw behavior is configurable between "Indent only selected"(`CombineIndent`) and "indent all"(`Static`)
 
 ## 0.1.0